refactor(desktop): extract page context helpers
